requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
braceexpand==0.1.7
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cmake==4.1.0
contourpy==1.3.3
cycler==0.12.1
datasets==4.1.1
deepspeed==0.15.4
dill==0.4.0
easydict==1.13
einops==0.8.1
fastapi==0.117.1
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.34.0
gradio_client==1.10.3
groovy==0.1.2
grpcio==1.75.0
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
kiwisolver==1.4.9
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
ninja==1.13.0
numpy==1.26.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.2.1
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
ruff==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.38.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.66.5
transformers==4.51.3
triton==3.4.0
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
wandb==0.22.0
webdataset==1.0.2
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
